pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codewithbrain"
version = "0.1.0"
dependencies = [
    "Django>=5.0.1",
    "gunicorn>=21.2.0",
    "whitenoise>=6.6.0",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
    "dj-database-url>=2.1.0",
    "djangorestframework>=3.14.0",
    "django-taggit>=5.0.1",
    "django-ckeditor-5>=0.2.10",
    "django-unfold>=0.22.0",
    "Pillow>=10.2.0",
    "python-slugify>=8.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-django>=4.8.0",
    "django-debug-toolbar>=4.2.0",
    "django-extensions>=3.2.3",
    "flake8>=7.0.0",
    "black>=24.1.1",
]